refactor: move JakartaPersistenceExtension to sql.repository package and update references
