+++ jnosql-jakarta-persistence/jnosql-jakarta-persistence-driver/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -12,4 +12,4 @@
 #
 #   Ondro Mihalyi
 #
-org.eclipse.jnosql.extensions.sql.repository.spi.JakartaPersistenceExtension
+org.eclipse.jnosql.extensions.sql.repository.JakartaPersistenceExtension
