+++ jnosql-jakarta-persistence/jnosql-jakarta-persistence-driver/src/test/java/ee/omnifish/jnosql/jakartapersistence/TestSupport.java
@@ -20,7 +20,7 @@ import jakarta.enterprise.inject.se.SeContainerInitializer;
 import org.eclipse.jnosql.jakartapersistence.communication.PersistenceDatabaseManager;
 import org.eclipse.jnosql.jakartapersistence.mapping.PersistenceDocumentTemplate;
 import org.eclipse.jnosql.jakartapersistence.mapping.cache.PersistenceUnitCacheProvider;
-import org.eclipse.jnosql.extensions.sql.repository.spi.JakartaPersistenceExtension;
+import org.eclipse.jnosql.extensions.sql.repository.JakartaPersistenceExtension;
 
 /**
  *
